- Move Furniture4YouApp to separate composable
