+++ app/build.gradle
@@ -60,7 +60,7 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.13.1'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.8.1'
     implementation 'androidx.activity:activity-compose:1.9.0'
-    implementation platform('androidx.compose:compose-bom:2025.01.00')
+    implementation platform('androidx.compose:compose-bom:2024.05.00')
     implementation 'androidx.compose.ui:ui'
     implementation 'androidx.compose.ui:ui-graphics'
     implementation 'androidx.compose.ui:ui-tooling-preview'
@@ -82,6 +82,7 @@ dependencies {
     // Dependency injection
     implementation 'com.google.dagger:hilt-android:2.51.1'
     implementation 'androidx.appcompat:appcompat:1.7.0'
+    debugImplementation 'androidx.compose.ui:ui-tooling:1.7.7'
     kapt 'com.google.dagger:hilt-compiler:2.51.1'
     implementation("androidx.hilt:hilt-navigation-compose:1.2.0")
